chore: credentials view UI restructure
